first attempt: moved everything to utils.py
